Introduce ContentSource and ContentLibraryProvider CRDs
In order to make virtual machine images available to a developer, a VI
admin exposes them by adding them to a content library and configuring
a cluster to use that content library. This configuration is done by
plugging in the content library uuid in VM operator's provider
ConfigMap.
This presents a few challenges:
- only one content library can be associated with a VM operator
instance
- image discovery has to be driven via the images controller. While
it works, this should be driven from top down. A content library
should create the images in that library

This change introduces new CRDs to enhance the image support in VM
operator.

*ContentSource*
A ContentSource describes a content source. This resources contains a
reference to a resource which describes a content provider. A content
library is one such example of a content provider. Content providers
have their owner references set to the ContentSource.

** Example ContentSource resource **
apiVersion: vmoperator.vmware.com/v1alpha1
kind: ContentSource
metadata:
  name: parunesh-contentsource
spec:
  providerRef:
    apiVersion: vmoperator.vmware.com/v1alpha1
    kind: ContentLibraryProvider
    name: parunesh-contentlib-vmtx
----

*ContentLibraryProvider*
A ContentLibraryProvider is a content provider which uniquely describes a
vSphere content library. Currently, only the content library UUID is
used to identify a library on vSphere.

** Example ContentLibraryProvider resource **
apiVersion: vmoperator.vmware.com/v1alpha1
kind: ContentLibraryProvider
metadata:
  name: parunesh-contentlib-vmtx
  ownerReferences:
  - apiVersion: vmoperator.vmware.com/v1alpha1
    kind: ContentSource
    name: parunesh-contentsource
    uid: 7bbd5927-fd57-4e8c-adae-8fa162cf80e8
spec:
  uuid: e5d287d1-a6e0-48fb-b6f0-34ca0996729b
----

VirtualMachineImage is modified to support both the old and new modes
of image discovery. This is done by listing all the images available
in content providers pointed by the installed ContentSources. This
change also adds integration tests to test the new image discovery
via CRDs.