Introduce OvfEnv VmMetadataTransport (#16)
OvfEnv is a standard way to transport key/value vm customization data from the infrastructure to the VM, as well as exposing them to guest os/apps within the VM. Possible key/value properties that are userConfigurable would be already defined in the .ovf image as defined in the OVF Spec 2.1.1 Section 11 https://www.dmtf.org/sites/default/files/standards/documents/DSP0243_2.1.1.pdf

This MR makes it possible to supply key/value fields to set the userConfigurable property fields on a .ovf during the VirtualMachine create time. The VirtualMachine spec already has a VmMetadata field which already supports a "ExtraConfig" transport, and this change adds another possible transport which is "OvfEnv".