Add Conditions to VirtualMachine status (#20)
To report VM status, we currently use Phase and PowerState. Phase is a high level abstraction of VM's execution phase, and PowerState, as implied by its name conveys VM power-state. We currently do not have a means to convey any fine-grained status that is actionable from user/client's perspective. We do have Events, but they are generally meant to be used as an equivalent of log messages for human consumption and not intended for event-driven automation.

It would be useful to have an equivalent of Conditions array present in the core Kubernetes resources like Pods and Nodes to report fine-grained status information that is actionable from the client's perspective.