Add VirtualMachineDisk, SystemDisk to spec
Currently, only VirtualMachineVolumes can be specified in the
VM service. These are limited to CNS volumes and are to be
use by PVCSI addons in TKG clusters.

This change introduces disk support to the VM service. A
barebones VirtualMachineDisk struct is added and can be specified
as a SystemDisk for a VirtualMachine. Future commits may
introduce AttachedDisks to the VirtualMachineSpec to specify
desired attachment for non-system disks.