Introduce VirtualMachineClassBinding CRD
The VirtualMachineClassBinding CRD represents a binding between a supervisor namespace and a VirtualMachineClass.

A VirtualMachineClass is a policy and configuration resource defining a set of attributes to configure of a VirtualMachine instance.

The purpose of this Binding CRD is to create a Namespace association
since the VirtualMachineClass is a cluster scoped resource.

For instance,
If a supervisor namespace has 3 VirtualMachineClassBinding resources with classRefs pointing to 3 VirtualMachineClasses,
then that namespace has access to three VM classes (small, medium, large).