Add default volume provisioning options for VirtualMachine
This change introduces additional configuration for VirtualMachine
that will allow users to specify what type of provisioning (e.g
thin or eager zeroed) for all the disks belonging to this
VirtualMachine. In a subsequent change, this configuration shall
also be available for each VirtualMachineVolume.

As part of this change, this configuration is added under a new
AdvancedOptions section of the VirtualMachine spec which was
planning to be introduced soon for such optional, advanced
configurations.

Testing done:
- make all