Add generic Conditions derived from cluster-api
This does replace VirtualMachineCondition with Condition structure.
We haven't started to set conditions yet, and this doesn't change
the REST API so I don't think it is really breaking, but we could
probably modify GetConditions/SetConditions to convert between the
types if desired.

The GetConditions/SetConditions methods are so the VirtualMachine
satisfies the special Patch helper Conditions support.