Change GuestOSCustomizable to SupportedGuestOS
The GuestOSDescriptor ids fetched from the cluster
configoptions don't indicate whether the osType is customizable by GOSC or not.
It is a cautionary warning for GuestOS compatibility with the esx hosts of the
cluster.

It makes sense to change the name to SupportedGuestOS
since it will be used to mark osTypes that are considered "unsupported"
as false.