Add Condition and Reasons for ContentSourceBindings and VM Images
In order to provide better feedback to users when a VirtualMachineImage
specified in the VirtualMachine yaml is not available on the API server, or, if
the  developer's namespace does not have access to those images, this change
introduces Reasons that can be set on the conditions as part of the
VirtualMachine's Status.

Following Reasons are defined:
- ContentSourceBindingNotFoundReason
If the developer's namespace does not have access to use the specified
VirtualMachineImage.

- VirtualMachineImageNotFoundReason
If the specified VirtualMachineImage is not found.

These reasons are set on the VirtualMachinePreReqReady condition.