Enforce enum restrictions on VM power state and metadata transport
The power state of the VirtualMachine from the spec should match exactly as the
power states defined in the vim 25 API descriptions. The only allowed values are
"powredOn" and "poweredOff".

Similarly, the supported values for VM metadata transport are OvfEnv and
ExtraConfig, but we do not enforce them.

This change enforces these values by leveraging the kubebuilder validation
markers.