Add missing field descriptions and various diff reduce from Service (#57)
Add missing descriptions to ClusterIP and ExternalName fields in the
Spec, so that a description is available via kubectl explains. Add some
missing +optional and omitempty, but there is some more validation we
should be doing before we can apply this more broadly.

Sync descriptions of various other fields from their Service counterparts.