Add ClusterMoID to the ClusterModuleStatus
This field will contain the domain-XXX that corresponds to the
ClusterComputeResource. The VC Cluster Modules API is not very
friendly for our needs: the module create API does not take a
name or other identifier, so we're force to keep the module ID
in our status to be able to look it up. With future HA support,
we need create the modules on each cluster, so we also need the
cluster in the status.