VMSVC-436:Remove VirtualMachineClass and ContentSource fields from the `kubectl get` output
Relevant slack thread: https://vmware.slack.com/archives/C013QBGESFR/p1607197143196200

This came up in a slack thread that having a `VirtualMachineClass` column in the output of `kubectl get vmclassbindings` is confusing since as of now, the bindings have the same name as the VM class names. The suggestion is to remove the additional column from the kubectl output.

This applies to the `kubectl get contentsourcebindings` as well.