Introduce annotation to pause a VM reconcile
In order to support VADP based backup/restore of VM Service VMs, we
need to pause the reconcile while the VM is recreated from backup on
the infrastructure. This change introduces a pause annotation to
achieve this. When set, VM operator will not reconcile this VM until
this annotation is removed.