VirtualMachinePublishRequest API
This patch introduces the new VirtualMachinePublishRequest API
for supporting self-service publication of VM Service VMs as new
VM Images in the same namespace as the VM being published.

Please note this patch also updates the Kubernetes dependencies
to v0.23.1 in order to leverage the metav1.Condition type from
the Kubernetes API machinery project.

The Go module version of Go is also updated to v1.17 since v1.13
is getting rather long in the tooth.